Port to Python
This might seem like an odd change, but I'm more familiar
with Python and don't see a massive upgrade for me learning/building
it in Go right now. I may re-write this in Go in the future as having
it as an executable would be great, or might consider another language.
But for now, being a fairly simple program, I think is fine to keep as
Python.