Refactor copy_files function and fold into sync
The sync() function wasn't really doing anything, and
was meant to the actual CLI `sync` command. This commit
adds the copy_files functionality into the sync command,
and also adds sha256 hashes to check whether or not the
file exists or is already in sync in the current directory.

This is so when `sync` is run it doesn't always have to copy
the files over; instead it looks to see if it even needs to
first.