Add newline to end of default yeo.json file
When the `init` command creates a yeo.json file,
it did not initally add a newline after creating
the JSON pbject and writing it to the file.

This change adds a newline character, so when the
file is printed to the terminal, the prompt appears
in the correct and expected place after running a
command on POSIX terminals.